requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu101
torchvision==0.8.2+cu101
pandas==1.3.3
numpy==1.21.2
opencv-python==4.5.3.56
scikit-learn==1.0
albumentations==1.0.3
Pillow==8.3.2
pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
tqdm==4.62.3
plotly==5.3.1
PyYAML==5.4.1

wandb==0.12.2
torchio==0.18.56
sacred==0.8.2
python-telegram-bot==13.7
